[development/pokemon] Fixing bug in global scratch and temp allocator where the max_size would return max of s32 while in fact it should return the max size that's supported of the underlying allocator

+++ source/1_engine/rex_engine/include/public/rex_engine/memory/global_allocators/global_scratch_allocator.h
@@ -7,10 +7,10 @@ namespace rex
   class GlobalScratchAllocator
   {
   public:
-    void* allocate(const s32 count);    // deallocates the storage reference by the pointer p.
-    void deallocate(void* const ptr, s32 count);
+    void* allocate(const s64 count);    // deallocates the storage reference by the pointer p.
+    void deallocate(void* const ptr, s64 count);
 
-    s32 max_size() const;
+    s64 max_size() const;
 
     template <typename U, typename... Args>
     void construct(U* p, Args&&... args)
